segmented renderAttendanceView
